Put the benchmark into CLI add to cabal, modify prepare to accept source positions as well as ranges.
