+++ src/cli/haskell-tools-cli.cabal
@@ -49,3 +49,18 @@ test-suite haskell-tools-cli-tests
                      , bytestring                >= 0.10 && < 0.11
   default-language:    Haskell2010
 
+benchmark cli-benchmark
+  type:                exitcode-stdio-1.0
+  build-depends:       base                      >= 4.9 && < 4.10
+                     , haskell-tools-cli         >= 0.3 && < 0.4
+                     , criterion                 >= 1.1 && < 1.2
+                     , time                      >= 1.6 && < 1.7
+                     , aeson                     >= 0.11 && < 0.12
+                     , directory                 >= 1.2 && < 1.3
+                     , filepath                  >= 1.4 && < 2.0
+                     , knob                      >= 0.1 && < 0.2
+                     , bytestring                >= 0.10 && < 0.11
+  hs-source-dirs:      benchmark
+  main-is:             Main.hs  
+
+
